Refactor(CA6): replace ca4 -> ca6
