+++ CA6/.gitignore
@@ -23,7 +23,7 @@ frontend/node_modules/
 # Logs and reports
 /logs/
 /test-output/
-/IE-CA4-logs.txt
+/IE-CA6-logs.txt
 
 # Miscellaneous
 *.iml
