+++ CA6/backend/pom.xml
@@ -11,12 +11,12 @@
 		<version>3.2.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-  <groupId>com.github.mohajel.IE.CA4</groupId>
-  <artifactId>CA4</artifactId>
+  <groupId>com.github.mohajel.IE.CA6</groupId>
+  <artifactId>CA6</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>war</packaging>
 
-  <name>CA4 Maven Webapp</name>
+  <name>CA6 Maven Webapp</name>
   <!-- FIXME change it to the project's website -->
   <url>http://www.example.com</url>
 
@@ -26,7 +26,7 @@
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
     <java.version>17</java.version>
-    <start-class>com.github.mohajel.IE.CA4.controllers.DemoApplication</start-class>
+    <start-class>com.github.mohajel.IE.CA6.controllers.DemoApplication</start-class>
 
   </properties>
 
@@ -115,7 +115,7 @@
   </dependencies>
 
   <build>
-    <finalName>CA4</finalName>
+    <finalName>CA6</finalName>
     <pluginManagement><!-- lock down plugins versions to avoid using Maven defaults (may be moved to
       parent pom) -->
       <plugins>
@@ -124,7 +124,7 @@
           <artifactId>exec-maven-plugin</artifactId>
           <version>3.4.1</version>
           <configuration>
-              <mainClass>com.github.mohajel.IE.CA4.Main</mainClass> <!-- Replace with your actual fully qualified class name that contains the main method -->
+              <mainClass>com.github.mohajel.IE.CA6.Main</mainClass> <!-- Replace with your actual fully qualified class name that contains the main method -->
           </configuration>
         </plugin>
         <plugin>
