--- CA6/backend/src/main/java/com/github/mohajel/IE/CA4/utils/inputForInit.txt
@@ -1,17 +1,0 @@
-reserveTable {"username": "Mostafa_Ebrahimi", "restaurantName": "The Commoner", "tableNumber": 1, "datetime": "2025-02-27 09:00"}
-reserveTable {"username": "Mohammad_Sadeghi", "restaurantName": "The Commoner", "tableNumber": 2, "datetime": "2025-02-27 11:00"}
-reserveTable {"username": "Mostafa_Ebrahimi", "restaurantName": "The Commoner", "tableNumber": 3, "datetime": "2025-02-28 15:00"}
-reserveTable {"username": "Mostafa_Ebrahimi", "restaurantName": "Sullivan's Steakhouse", "tableNumber": 1, "datetime": "2025-02-27 18:00"}
-reserveTable {"username": "Mohammad_Sadeghi", "restaurantName": "Sullivan's Steakhouse", "tableNumber": 2, "datetime": "2025-02-27 13:00"}
-reserveTable {"username": "Mostafa_Ebrahimi", "restaurantName": "Eddie Merlot's Prime Aged Beef & Seafood", "tableNumber": 1, "datetime": "2025-02-27 13:00"}
-reserveTable {"username": "Mohammad_Sadeghi", "restaurantName": "Eddie Merlot's Prime Aged Beef & Seafood", "tableNumber": 2, "datetime": "2025-02-27 18:00"}
-reserveTable {"username": "Mostafa_Ebrahimi", "restaurantName": "Eddie Merlot's Prime Aged Beef & Seafood", "tableNumber": 1, "datetime": "2025-02-28 16:00"}
-reserveTable {"username": "Mohammad_Sadeghi", "restaurantName": "The Winery Restaurant & Wine Bar", "tableNumber": 1, "datetime": "2025-02-27 16:00"}
-reserveTable {"username": "Mostafa_Ebrahimi", "restaurantName": "The Winery Restaurant & Wine Bar", "tableNumber": 2, "datetime": "2025-02-27 19:00"}
-reserveTable {"username": "Mohammad_Sadeghi", "restaurantName": "The Winery Restaurant & Wine Bar", "tableNumber": 1, "datetime": "2025-02-28 15:00"}
-reserveTable {"username": "Mostafa_Ebrahimi", "restaurantName": "Et Voilà!", "tableNumber": 1, "datetime": "2025-02-27 17:00"}
-reserveTable {"username": "Mohammad_Sadeghi", "restaurantName": "Et Voilà!", "tableNumber": 2, "datetime": "2025-02-27 18:00"}
-reserveTable {"username": "Mostafa_Ebrahimi", "restaurantName": "North Italia", "tableNumber": 1, "datetime": "2025-02-27 13:00"}
-reserveTable {"username": "Mohammad_Sadeghi", "restaurantName": "North Italia", "tableNumber": 2, "datetime": "2025-02-27 19:00"}
-reserveTable {"username": "Mostafa_Ebrahimi", "restaurantName": "KUJIRA", "tableNumber": 1, "datetime": "2025-02-27 22:00"}
-reserveTable {"username": "Mohammad_Sadeghi", "restaurantName": "KUJIRA", "tableNumber": 2, "datetime": "2025-02-27 03:00"}
