+++ CA6/backend/src/main/java/com/github/mohajel/IE/CA6/controllers/ReviewController.java
@@ -1,6 +1,5 @@
-package com.github.mohajel.IE.CA4.controllers;
+package com.github.mohajel.IE.CA6.controllers;
 
-import com.github.mohajel.IE.CA4.MizdooniApp;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.http.ResponseEntity;
@@ -8,6 +7,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.github.mohajel.IE.CA6.MizdooniApp;
 
 @RestController
 @RequestMapping("/reviews")
