+++ CA6/backend/src/main/java/com/github/mohajel/IE/CA6/models/Address.java
@@ -1,4 +1,4 @@
-package com.github.mohajel.IE.CA4.models;
+package com.github.mohajel.IE.CA6.models;
 
 import org.json.JSONObject;
