+++ CA6/backend/src/main/java/com/github/mohajel/IE/CA6/models/Review.java
@@ -1,8 +1,8 @@
-package com.github.mohajel.IE.CA4.models;
+package com.github.mohajel.IE.CA6.models;
 
 import org.json.JSONObject;
 
-import com.github.mohajel.IE.CA4.utils.MizdooniError;
+import com.github.mohajel.IE.CA6.utils.MizdooniError;
 
 public class Review {
